feat(formulas): Display the first formula upon screen load
This commit makes a change so that the first formula is displayed on the
right when the screen is loaded. The purpose of this change is to enhance
the user experience. By showing the first formula by default, we eliminate
the need for users to manually select a formula

Jira Issue: DS-536